grpcio/pyproject.toml
[project]

name = "grpcio_app"
version = "1.0.0"
description = ""
authors = [{name = "Flet Developer", email = "your@email.com"}]
dependencies = ["flet==0.25.0.dev3750", "grpcio", "protobuf"]

[tool.flet]
android.adaptive_icon_background = "#ffffff"